Various refactorings
