+++ src/yukatool/Tasks/UnpackTask.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using Yuka.Data;
 using Yuka.Script;
+using Yuka.Data.Factory;
 
 namespace Yuka.Tasks {
 	class UnpackTask : Task {
